requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python-headless>=4.5.3
Pillow>=9.0.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Web interface dependencies
streamlit>=1.25.0
plotly>=5.10.0

# Image processing
scikit-image>=0.19.0

# Document generation
python-docx>=0.8.11

# Visualization
seaborn>=0.11.2

# Deep Learning (for MMFusion)
torch>=1.9.0
torchvision>=0.10.0
